pyproject.toml
[build-system]
requires = [
    "hatchling",
	"hatch-vcs"
]
build-backend = "hatchling.build"

[project]
name = "newport-esa"
version = "0.0.1"
authors = [
  { name="John Houlihan", email="houlihaj@uci.edu" },
]
description = "A Python API for the Newport Model ESA-C Ultra-Resolution Actuator Controller."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
]
dependencies = [
    "PyVISA-py >= 0.7.2",
]

[project.urls]
Homepage = "https://github.com/houlihaj/newport-esa"
Issues = "https://github.com/houlihaj/newport-esa/issues"

# https://hatch.pypa.io/latest/config/metadata/
[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.sdist]
include = ["/src", "/tests"]

[tool.coverage.run]
source = ["src/newport_esa"]

[tool.hatch.build.targets.wheel]
packages = ["src/newport_esa"]
